api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.32.3
redis==5.0.1
meilisearch==0.28.3
python-json-logger==2.0.7
prometheus-client==0.19.0
flashtext
openai
sentence-transformers
transformers==4.37.2
google-generativeai

pymupdf
python-docx
python-multipart
pgvector
uszipcode
alembic
numpy<2.0.0

